+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## unreleased
 
+## 1.0.0
+
+* [THORN-17] 1.0.0 release - all MCCS repos
+
+## 0.10.0
+
+* [THORN-12] Add methods/attributes to measure data rate over receiver interface.
+
+## 0.9.2
+
+* [SKB-705] DaqHandler to handle DAQ receiver restart gracefully while bandpass monitoring is active
+* [THORN-35] Update CODEOWNERS.
+
+## 0.9.1
+
+* [SKB-610] Add StationID as a configuration option.
+
 ## 0.9.0
 
 * [MCCS-2227] support SIGTERM in daq receivers so that Pods terminate gracefully